refactor: transaction 범위를 줄이기 위해 db 저장 메소드 분리
