+++ infra/src/main/java/com/fintory/infra/domain/stock/service/overseas/OverseasStockRankServiceImpl.java
@@ -4,18 +4,20 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fintory.common.exception.DomainErrorCode;
 import com.fintory.common.exception.DomainException;
-import com.fintory.domain.stock.dto.overseas.core.OverseasStockRankData;
 import com.fintory.domain.stock.dto.overseas.wrapper.OverseasStockRankDataWrapper;
 import com.fintory.domain.stock.model.Stock;
 import com.fintory.domain.stock.model.StockRank;
 import com.fintory.domain.stock.service.overseas.OverseasStockRankService;
 import com.fintory.infra.domain.stock.repository.StockRankRepository;
 import com.fintory.infra.domain.stock.repository.StockRepository;
+import com.fintory.infra.domain.stock.service.overseas.saver.OverseasStockRankSaverService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.http.*;
+import org.springframework.retry.annotation.Backoff;
+import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.ResourceAccessException;
@@ -32,6 +34,7 @@ public class OverseasStockRankServiceImpl implements OverseasStockRankService {
 
     private final StockRepository stockRepository;
     private final StockRankRepository stockRankRepository;
+    private final OverseasStockRankSaverService overseasStockRankSaverService;
     private final RedisTemplate<Object, Object> redisTemplate;
     private final RestTemplate restTemplate;
     private final ObjectMapper objectMapper;
@@ -47,7 +50,7 @@ public class OverseasStockRankServiceImpl implements OverseasStockRankService {
 
     //NOTE 랭킹 데이터는 상대적 비교가 필요해서 전체적인 일관성이 필요함 -> 랭킹이 동일하더라도 프론트에서 받은 데이터를 정렬해서 표시
     @Override
-    @Transactional //NOTE 너무 큰 트랜잭션 -> 타임아웃 추가 가능성 열어두기
+    @Retryable(maxAttempts=3, backoff = @Backoff(delay = 1000))
     public void initiateOverseasStockRank(){
         List<Stock> stockList = stockRepository.findByCurrencyName("USD");
         String token = (String) redisTemplate.opsForValue().get("kis-access-token");
@@ -78,7 +81,8 @@ public class OverseasStockRankServiceImpl implements OverseasStockRankService {
     }
 
     //순위를 얻는데 필요한 데이터 조회
-    private void processStockRankData(String code, String token) {
+    @Transactional
+    public void processStockRankData(String code, String token) {
         try {
             String url = UriComponentsBuilder.fromHttpUrl(baseUrl)
                     .path("/uapi/overseas-price/v1/quotations/price-detail")
@@ -102,7 +106,7 @@ public class OverseasStockRankServiceImpl implements OverseasStockRankService {
 
             if (response.getStatusCode().is2xxSuccessful()) {
                 OverseasStockRankDataWrapper wrapper = objectMapper.readValue(response.getBody(), OverseasStockRankDataWrapper.class);
-                saveStockRankData(code, wrapper); //db에 데이터 저장
+                overseasStockRankSaverService.saveStockRankData(code, wrapper); //db에 데이터 저장
             } else {
                 log.error("순위 관련 데이터 조회 실패: {} - 응답이 비어있음", code);
                 throw new DomainException(DomainErrorCode.API_RESPONSE_EMPTY);
@@ -120,37 +124,6 @@ public class OverseasStockRankServiceImpl implements OverseasStockRankService {
             log.error("예상치 못한 오류: {} - {}", code, e.getMessage());
             throw new DomainException(DomainErrorCode.INTERNAL_SERVER_ERROR);
         }
-    }
-
-    //순위를 얻는데 필요한 데이터 저장 메서드
-    private void saveStockRankData(String code, OverseasStockRankDataWrapper response) {
-        if (response == null || response.output() == null) {
-            log.warn("순위 관련 데이터 응답이 비어있음: {}", code);
-            throw new DomainException(DomainErrorCode.API_RESPONSE_EMPTY);
-        }
-
-        OverseasStockRankData item = response.output();
-
-        if (item == null) {
-            log.warn("순위 관련 응답에서 데이터를 찾을 수 없음");
-            throw new DomainException(DomainErrorCode.STOCK_DATA_NOT_FOUND);
-        }
-
-        StockRank stockRank = stockRankRepository.findByStockCode(code).orElse(null);
-        Stock stock = stockRepository.findByCode(code).orElseThrow(() -> new DomainException(DomainErrorCode.STOCK_NOT_FOUND));
-
-        if (stockRank == null) {
-            stockRank = StockRank.builder()
-                    .tradingVolume(item.tradingVolume())
-                    .rocRate(item.roc())
-                    .marketCap(item.marketCap())
-                    .stock(stock)
-                    .build();
-        } else {
-            stockRank.updateStockRankData(item.marketCap(), item.roc(), item.tradingVolume());
-        }
-
-        stockRankRepository.save(stockRank);
     }
 
     //순위 데이터 생성 및 저장
